Cleaner setting of editor "enter" button listener
